Item Holding, Basic Shell, Animated Piranha
Built a shell that will launch when Mario bumps into it, but if he is holding the run button and its stationary he'll pick it up. I also modified his animation system to swap between holding models and regular models. Imported piranha models, wrote a basic script to animate it