Chuck, Mountains, and Bombs?
Finally imported chuck materials, created a blueprint and an enum for behavior states. Built functionality to jump and clap when mario is detected, and Jump High if it detects that Mario is above him.
Also got tired of the game being in a massive open field so I made some of the Map assets to fill in the blanks. On the one hand, The oversized blocks look a bit out of place, but on the other hand its much better than nothing.
Also created some bomb folders.