Game Framework and HUD
Began creating files for unreals Game Framework., beginning with the HUD.

Renamed MarioMode to LevelMode as it will be the GameMode used In levels, I'll also need to make a MapMode and potentially a MenuMode.

Created Scripts for LevelHud and Created HUD BP that inherits from it. also created a Lives widget.